Move desktop Pi services behind runtime host
